pyproject.toml
[tool.poetry]
name = "nomina-py"
version = "0.1.0"
description = "Plataforma Web para Administrar Nómina de empresa TriMark cadena de supermercados."
authors = ["Tomás Fuentes Milla <tomas.fuentes07@inacapmail.cl>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
Django = "^4.2.5"
django-session-timeout = "^0.1.0"
pymongo = "^4.6.1"
python-dotenv = "^1.0.0"
py-bcrypt = "^0.4"
pytz = "^2023.3.post1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
